pyproject.toml
[tool.poetry]
name = "datacamp-scraper"
version = "0.1.0"
description = "A web crawler to crawl and scrape datacamp courses and download course materials"
authors = ["rifatrakib <Rakib.1508@outlook.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "datacamp_scraper"}]

[tool.poetry.dependencies]
python = "^3.9"
Scrapy = "^2.7.1"
pymongo = "^4.3.2"
python-dotenv = "^0.21.0"
types-requests = "^2.28.11.4"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
isort = "^5.10.1"
mypy = "^0.990"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.tox
    | venv
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 119
